+++ src/mmSolver/sfm/camera_relative_pose.h
@@ -70,7 +70,7 @@ bool robust_relative_pose(const openMVG::cameras::IntrinsicBase *intrinsics1,
                           openMVG::sfm::RelativePose_Info &relativePose_info,
                           const std::pair<size_t, size_t> &size_ima1,
                           const std::pair<size_t, size_t> &size_ima2,
-                          const size_t max_iteration_count);
+                          const uint32_t max_iteration_count);
 
 bool compute_relative_pose(
     const int32_t image_width_a, const int32_t image_width_b,
